add useColumn hooks to specialistlist, tagslist
