+++ packages/webapp/src/components/lists/SpecialistList/index.tsx
@@ -5,23 +5,22 @@
 
 import styles from './index.module.scss'
 import type { StandardFC } from '~types/StandardFC'
-import { RoleType, User } from '@cbosuite/schema/dist/client-types'
+import { User } from '@cbosuite/schema/dist/client-types'
 import cx from 'classnames'
-import { MultiActionButton, IMultiActionButtons } from '~components/ui/MultiActionButton2'
+import { IMultiActionButtons } from '~components/ui/MultiActionButton2'
 import { useWindowSize } from '~hooks/useWindowSize'
 import { useState } from 'react'
 import { useBoolean } from '@fluentui/react-hooks'
 import { Panel } from '~ui/Panel'
 import { AddSpecialistForm } from '~components/forms/AddSpecialistForm'
 import { EditSpecialistForm } from '~components/forms/EditSpecialistForm'
-import { PaginatedList, IPaginatedListColumn } from '~components/ui/PaginatedList'
+import { PaginatedList } from '~components/ui/PaginatedList'
 import { useSpecialist } from '~hooks/api/useSpecialist'
 import { useTranslation } from '~hooks/useTranslation'
 import { wrap } from '~utils/appinsights'
 import { useCurrentUser } from '~hooks/api/useCurrentUser'
 import { useUserSearchHandler } from '~hooks/useUserSearchHandler'
-import { SpecialistTitleColumnItem } from '~components/ui/SpecialistTitleColumnItem'
-import { SpecialistMobileCard } from '~components/ui/SpecialistMobileCard'
+import { useMobileColumns, usePageColumns } from './columns'
 
 interface SpecialistListProps {
 	title?: string
@@ -49,7 +48,7 @@ export const SpecialistList: StandardFC<SpecialistListProps> = wrap(function Spe
 
 	const searchList = useUserSearchHandler(specialistList, setFilteredList)
 
-	const columnActionButtons: IMultiActionButtons<User>[] = isAdmin
+	const actions: IMultiActionButtons<User>[] = isAdmin
 		? [
 				{
 					name: t('specialistListRowActions.edit'),
@@ -62,65 +61,8 @@ export const SpecialistList: StandardFC<SpecialistListProps> = wrap(function Spe
 		  ]
 		: []
 
-	const pageColumns: IPaginatedListColumn[] = [
-		{
-			key: 'name',
-			name: t('specialistListColumns.name'),
-			onRenderColumnItem(user: User) {
-				return <SpecialistTitleColumnItem user={user} />
-			}
-		},
-		{
-			key: 'numOfEngagement',
-			name: t('specialistListColumns.numOfEngagement'),
-			onRenderColumnItem(user: User) {
-				return (
-					<span>
-						{user?.engagementCounts?.active || 0} {t('specialistStatus.assigned')},{' '}
-						{user?.engagementCounts?.closed || 0} {t('specialistStatus.closed')}
-					</span>
-				)
-			}
-		},
-		{
-			key: 'userName',
-			name: t('specialistListColumns.username'),
-			onRenderColumnItem(user: User) {
-				return `@${user.userName}`
-			}
-		},
-		{
-			key: 'permissions',
-			name: t('specialistListColumns.permissions'),
-			onRenderColumnItem(user: User) {
-				return (
-					<>
-						{user?.roles.filter((r) => r.roleType === RoleType.Admin).length > 0
-							? t('specialistRoles.admin')
-							: t('specialistRoles.user')}
-					</>
-				)
-			}
-		},
-		{
-			key: 'actionColumn',
-			name: '',
-			className: 'w-100 d-flex justify-content-end',
-			onRenderColumnItem(user: User) {
-				return <MultiActionButton columnItem={user} buttonGroup={columnActionButtons} />
-			}
-		}
-	]
-
-	const mobileColumn: IPaginatedListColumn[] = [
-		{
-			key: 'cardItem',
-			name: 'cardItem',
-			onRenderColumnItem(user: User) {
-				return <SpecialistMobileCard user={user} actions={columnActionButtons} />
-			}
-		}
-	]
+	const pageColumns = usePageColumns(actions)
+	const mobileColumns = useMobileColumns(actions)
 
 	return (
 		<div className={cx('mt-5 mb-5', styles.specialistList, 'specialistList')}>
@@ -129,7 +71,7 @@ export const SpecialistList: StandardFC<SpecialistListProps> = wrap(function Spe
 				hideListHeaders={!isMD}
 				list={filteredList}
 				itemsPerPage={isMD ? 20 : 10}
-				columns={isMD ? pageColumns : mobileColumn}
+				columns={isMD ? pageColumns : mobileColumns}
 				rowClassName='align-items-center'
 				addButtonName={t('specialistAddButton')}
 				onSearchValueChange={searchList}
